modified:   library/libsoft0/linux/linux.memory.file.c 	modified:   library/libsoft0/win/win.memory.file.c 	modified:   library/libsoft1/libnet/manager/net.server.c 	modified:   library/libsoft1/libnet/worker/net.http.c 	modified:   project/none+arm+linux+cli/makefile 	modified:   project/none+arm+linux+fb/makefile 	modified:   project/none+arm+linux+tui/makefile 	modified:   project/none+arm+linux+xlib/makefile

+++ library/libsoft1/libnet/worker/net.http.c
@@ -13,8 +13,8 @@ int findtail(char* p);
 int cmp(u8*,u8*);
 int ncmp(u8*,u8*,int);
 //
-int readfile(char* name, char* mem, int offset, int count);
-int writefile(char* name, char* mem, int offset, int count);
+int readfile(char* name, char* mem, u64 offset, u64 count);
+int writefile(char* name, char* mem, u64 offset, u64 count);
 int readserver(u64 fd, u8* addr, u64 offset, u64 count);
 int writeserver(u64 fd, u8* addr, u64 offset, u64 count);
 int notify_delete(u64 x);
